Add SQLColumnPropertyType
