+++ Sources/SQLKitTyping/Macros.swift
@@ -10,3 +10,6 @@ public macro Column(namespace: String) = #externalMacro(module: "SQLKitTypingMac
 
 @attached(accessor)
 public macro EraseProperty() = #externalMacro(module: "SQLKitTypingMacros", type: "EraseProperty")
+
+@freestanding(declaration, names: arbitrary)
+public macro SQLColumnPropertyType(name: String, type: any Decodable.Type) = #externalMacro(module: "SQLKitTypingMacros", type: "SQLColumnPropertyType")
