+++ Sources/SQLKitTyping/PropertyBuilder/SQLTypedQueryFetcher.swift
@@ -1,0 +1,51 @@
+import SQLKit
+
+public protocol SQLTypedQueryFetcher<Row>: SQLQueryFetcher {
+    associatedtype Row: Decodable
+}
+
+extension SQLTypedQueryFetcher {
+    public func first(
+        prefix: String? = nil,
+        keyDecodingStrategy: SQLRowDecoder.KeyDecodingStrategy = .useDefaultKeys,
+        userInfo: [CodingUserInfoKey: any Sendable] = [:]
+    ) async throws -> Row? {
+        guard let row = try await first() as (any SQLRow)? else {
+            return nil
+        }
+        let decoder = SQLRowDecoder(prefix: prefix, keyDecodingStrategy: keyDecodingStrategy, userInfo: userInfo)
+        return try row.decode(model: Row.self, with: decoder)
+    }
+
+    public func firstWithRow(
+        prefix: String? = nil,
+        keyDecodingStrategy: SQLRowDecoder.KeyDecodingStrategy = .useDefaultKeys,
+        userInfo: [CodingUserInfoKey: any Sendable] = [:]
+    ) async throws -> (Row, any SQLRow)? {
+        guard let row = try await first() as (any SQLRow)? else {
+            return nil
+        }
+        let decoder = SQLRowDecoder(prefix: prefix, keyDecodingStrategy: keyDecodingStrategy, userInfo: userInfo)
+        return (try row.decode(model: Row.self, with: decoder), row)
+    }
+
+    public func all(
+        prefix: String? = nil,
+        keyDecodingStrategy: SQLRowDecoder.KeyDecodingStrategy = .useDefaultKeys,
+        userInfo: [CodingUserInfoKey: any Sendable] = [:]
+    ) async throws -> [Row] {
+        let rows = try await all() as [any SQLRow]
+        let decoder = SQLRowDecoder(prefix: prefix, keyDecodingStrategy: keyDecodingStrategy, userInfo: userInfo)
+        return try rows.map { try $0.decode(model: Row.self, with: decoder) }
+    }
+
+    public func allWithRow(
+        prefix: String? = nil,
+        keyDecodingStrategy: SQLRowDecoder.KeyDecodingStrategy = .useDefaultKeys,
+        userInfo: [CodingUserInfoKey: any Sendable] = [:]
+    ) async throws -> [(Row, any SQLRow)] {
+        let rows = try await all() as [any SQLRow]
+        let decoder = SQLRowDecoder(prefix: prefix, keyDecodingStrategy: keyDecodingStrategy, userInfo: userInfo)
+        return try rows.map { (try $0.decode(model: Row.self, with: decoder), $0) }
+    }
+}
